Refactor API console into REST console
